Handle rate limiting error when fetching
If CloudWatch raises a ThrottlingException it means there's too many
requests. This error bubbles up and causes the plugin to crash in
Logstash which then re-starts it. Instead we break for `interval`
duration and then start again from the log group where we left off.

After a log group is read we push its priority to bottom to ensure every
log group is inspected.